Add formula for FreeSWITCH compatible iLBC library
The canonical formula for libilbc homebrew/Library/Formula/libilbc.rb,
based on the recent iLBC library release, is API incompatible with the
iLBC API used in FreeSWITCH. The in-tree version of libs/ilbc in the
FreeSWITCH source tree is based on a different API. This formula makes
that version available for install via homebrew. Additionally, it
creates a pkgconfig file so that the FreeSWITCH configure checks can
pick up this version.

From a quick perusal of the git logs, it appears that the source code
for the codec has been unmodified in the FreeSWITCH source tree, so
this Formula is based on the original tarball. The changes in the
FreeSWITCH source tree seem to be around getting the windows build to
work (these changes are of no concern to the homebrew build), and
silencing coverity warnings (since we're now packaging iLBC outside
the source tree, even these should not be essential).